Uploading dataset GSE29008
Here the datasets for the HCT8 cells from GEO dataset GSE29008 are uploaded. The data is read into hct_eight. RMA normalization is performed on the data followed by converting it into an expression matrix. The column names are changed according to the conditions in which they were examined. here the effects of the toxins were examined at 2,6 and 24 hours