Uploading dataset GSE44091
The datasets for the mouse cells are uploaded here. There are 32 samples, obtained at 2,4 and 16 hours. The data is loaded, normalization is performed by RMA( Robust Multi- Array Average). The corresponding column names are provided based on the conditions in which they were examined. They are then converted into an expression matrix and the values in the table are log transformed for normal distribution.